+++ src/app/app/urls.py
@@ -26,8 +26,8 @@ urlpatterns = [
     # Deployed docs API.
     url(r'^api/docs/', include('docs.urls')),
 
-    # Feedback
-    url(r'^api/feedback/', include('feedback.urls')),
+    # stats
+    url(r'^api/stats/', include('stats.urls')),
 
     # Wagtail pages.
     url(r'^', include(wagtail_urls)),
